+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2024-10-21
+
+### Added
+
+- Introduced new `passes` method for custom filtering functions
+- Added `is_type` method for type checking in queries
+- Expanded test suite to cover new `passes` and `is_type` functionalities
+- Implements a blocking `finish_backup` method to databases to complete a backup before proceeding
+
 ## [1.1.0] - 2024-10-20
 
 ### Added